+++ runtimes/runtimes/chat/baseChat.ts
@@ -45,6 +45,12 @@ import {
     createPromptNotificationType,
     LogInlineChatResultParams,
     logInlineChatResultNotificationType,
+    listConversationsRequestType,
+    ListConversationsParams,
+    ListConversationsResult,
+    ConversationClickParams,
+    conversationClickRequestType,
+    ConversationClickResult,
 } from '../../protocol'
 import { Chat } from '../../server-interface'
 
@@ -131,5 +137,13 @@ export class BaseChat implements Chat {
 
     public onLogInlineChatResult(handler: NotificationHandler<LogInlineChatResultParams>) {
         this.connection.onNotification(logInlineChatResultNotificationType.method, handler)
+    }
+  
+    public onListConversations(handler: RequestHandler<ListConversationsParams, ListConversationsResult, void>) {
+        this.connection.onRequest(listConversationsRequestType.method, handler)
+    }
+
+    public onConversationClick(handler: RequestHandler<ConversationClickParams, ConversationClickResult, void>) {
+        this.connection.onRequest(conversationClickRequestType.method, handler)
     }
 }